Revert "Check whether the StoreId is inside the store, before doing anything on the FS"
This reverts commit 373502217e62ad07f76af4579680ad7156bbe390
as this commit was introducing a bug.

The StoreId type says `/test/example` for a store id path, which is
completely valid, as the root (`/`) is the store itself. The
id_in_store() function assumed that the store-id is the full
(file-system) path to the store entry, which is false.

This commit does not introduce a fix but revert the check.