Rewrite process_rw_result()
It did not cover the aspect that the passed toml `Value` object was an
Array in the first place and assumed that it is an `String` instead,
which caused the function always to return an Result::Err, which was
obviousely buggy.

This is almost a rewrite of the process_rw_result() function which takes
the additional layer (Array(String)) into account and therefor fixes
this issue.