Remove "caused by" marker
As we are in debug mode anyways, a developer using this should know by
the trace number that the next line is the cause.

As we use the debug!() macro here, this would be printed on a new line,
which is rather distracting instead of helpful. Therefor this patch
removes this line.