Make StoreIdIterator generic
This patch makes the StoreIdIterator type generic over all iterators
with Item = StoreId.

As the StoreIdIterator type was previousely used for iterating over a
glob() result, I had to wrap this result type in another iterator type
which is then wrapped in StoreIdIterator.

With this patch applied, other libraries may use the StoreIdIterator
type to abstract things away in their implementation.