Adds a Makefile for building modules
One can now type:

```
$> make
```

Which builds all the modules and places them in a `out/` directory of the project root for easy
finding.

Alternatively one can type:

```
$> make $module
```

Where `$module` is one of the `imag-$module` names, such as `counter`, `link`, etc.

To clean up the `out/` directory it's `$> make clean`