Fix check whether we should checkout the ensure_branch
We should test this not at the top of the function, as if we do not have
to check out the ensure_branch, we still have to verify that there
actually is a branch checked out we have to ensure.

What I'm saying is: We might have to be on "master", but we might not
have to check it out if we aren't.

Therefor, this commit moves the location where we check whether we
should checkout the branch.