Fix test backend implementation
We must return an error if the file is not found here.

This is because if we unwrap() here, we panic if the store action was
Store::retrieve() and there wasn't a file there. We then unwrap() on
None and panic because of this, causing all other tests to panic as
well because the Mutex gets corrupted.

The store handles FileNotFound errors on its own, so it is safe to
return the error here.