Adapt for new API of Habit::next_instance_date()
With cd8f0beae864c06f02b0d43fe4b63a0ba580229a
the interface of `HabitTemplate::next_instance_date()` changed to return
a `Result<Option<_>>` rather than a `Result<_>` because the
next_instance_date could be a None if the Habit was closed already.

This patch adapts imag-habit for this new interface.